Split ADContext off to a separate file.
