+++ lib/SILOptimizer/Mandatory/Differentiation.cpp
@@ -55,6 +55,7 @@
 #include "swift/SILOptimizer/Utils/CFGOptUtils.h"
 #include "swift/SILOptimizer/Utils/LoopUtils.h"
 #include "swift/SILOptimizer/Utils/SILOptFunctionBuilder.h"
+#include "swift/SILOptimizer/Utils/Differentiation/ADContext.h"
 #include "swift/SILOptimizer/Utils/Differentiation/Common.h"
 #include "swift/SILOptimizer/Utils/Differentiation/DerivativeLookup.h"
 #include "llvm/ADT/APSInt.h"
@@ -141,25 +142,6 @@ static CanType joinElementTypesFromValues(SILValueRange &&range,
   return TupleType::get(elts, ctx)->getCanonicalType();
 }
 
-/// Given an operator name, such as '+', and a protocol, returns the '+'
-/// operator. If the operator does not exist in the protocol, returns null.
-static FuncDecl *findOperatorDeclInProtocol(DeclName operatorName,
-                                            ProtocolDecl *protocol) {
-  assert(operatorName.isOperator());
-  // Find the operator requirement in the given protocol declaration.
-  auto opLookup = protocol->lookupDirect(operatorName);
-  for (auto *decl : opLookup) {
-    if (!decl->isProtocolRequirement())
-      continue;
-    auto *fd = dyn_cast<FuncDecl>(decl);
-    if (!fd || !fd->isStatic() || !fd->isOperator())
-      continue;
-    return fd;
-  }
-  // Not found.
-  return nullptr;
-}
-
 /// Returns the "constrained" derivative generic signature given:
 /// - An original SIL function type.
 /// - A wrt parameter index subset.
@@ -222,35 +204,11 @@ static GenericParamList *cloneGenericParameters(ASTContext &ctx,
   return GenericParamList::create(ctx, SourceLoc(), clonedParams, SourceLoc());
 }
 
-/// Given an `differentiable_function` instruction, finds the corresponding
-/// differential operator used in the AST. If no differential operator is found,
-/// return nullptr.
-static DifferentiableFunctionExpr *
-findDifferentialOperator(DifferentiableFunctionInst *inst) {
-  return inst->getLoc().getAsASTNode<DifferentiableFunctionExpr>();
-}
-
-/// Returns the underlying instruction for the given SILValue, if it exists,
-/// peering through function conversion instructions.
-template<class Inst>
-static Inst *peerThroughFunctionConversions(SILValue value) {
-  if (auto *inst = dyn_cast<Inst>(value))
-    return inst;
-  if (auto *thinToThick = dyn_cast<ThinToThickFunctionInst>(value))
-    return peerThroughFunctionConversions<Inst>(thinToThick->getOperand());
-  if (auto *convertFn = dyn_cast<ConvertFunctionInst>(value))
-    return peerThroughFunctionConversions<Inst>(convertFn->getOperand());
-  if (auto *partialApply = dyn_cast<PartialApplyInst>(value))
-    return peerThroughFunctionConversions<Inst>(partialApply->getCallee());
-  return nullptr;
-}
-
 //===----------------------------------------------------------------------===//
 // Auxiliary data structures
 //===----------------------------------------------------------------------===//
 
 namespace {
-class ADContext;
 
 /// Linear map struct and branching trace enum information for an original
 /// function and and derivative function (JVP or VJP).
@@ -628,251 +586,6 @@ public:
   }
 };
 
-/// Stores `apply` instruction information calculated by VJP generation.
-struct NestedApplyInfo {
-  /// The differentiation indices that are used to differentiate this `apply`
-  /// instruction.
-  SILAutoDiffIndices indices;
-  /// The original pullback type before reabstraction. `None` if the pullback
-  /// type is not reabstracted.
-  Optional<CanSILFunctionType> originalPullbackType;
-};
-
-static inline llvm::raw_ostream &operator<<(llvm::raw_ostream &os,
-                                            DifferentiationInvoker invoker) {
-  invoker.print(os);
-  return os;
-}
-
-//===----------------------------------------------------------------------===//
-// ADContext - Per-module contextual information for the Differentiation pass.
-//===----------------------------------------------------------------------===//
-
-class ADContext {
-private:
-  /// Reference to the main transform.
-  SILModuleTransform &transform;
-
-  /// The module where Differentiation is performed on.
-  SILModule &module;
-
-  /// AST context.
-  ASTContext &astCtx = module.getASTContext();
-
-  /// Shared pass manager.
-  SILPassManager &passManager;
-
-  /// The worklist (stack) of `differentiable_function` instructions to be
-  /// processed.
-  SmallVector<DifferentiableFunctionInst *, 32> differentiableFunctionInsts;
-
-  /// The set of `differentiable_function` instructions that have been
-  /// processed. Used to avoid reprocessing invalidated instructions.
-  /// NOTE(TF-784): if we use `CanonicalizeInstruction` subclass to replace
-  /// `ADContext::processDifferentiableFunctionInst`, this field may be removed.
-  SmallPtrSet<DifferentiableFunctionInst *, 32>
-      processedDifferentiableFunctionInsts;
-
-  /// Mapping from witnesses to invokers.
-  /// `SmallMapVector` is used for deterministic insertion order iteration.
-  SmallMapVector<SILDifferentiabilityWitness *, DifferentiationInvoker, 32>
-      invokers;
-
-  /// Mapping from `differentiable_function` instructions to result indices.
-  DenseMap<DifferentiableFunctionInst *, unsigned> resultIndices;
-
-  /// Mapping from original `apply` instructions to their corresponding
-  /// `NestedApplyInfo`s.
-  DenseMap<ApplyInst *, NestedApplyInfo> nestedApplyInfo;
-
-  /// List of generated functions (JVPs, VJPs, pullbacks, and thunks).
-  /// Saved for deletion during cleanup.
-  SmallVector<SILFunction *, 32> generatedFunctions;
-
-  /// List of references to generated functions.
-  /// Saved for deletion during cleanup.
-  SmallVector<SILValue, 32> generatedFunctionReferences;
-
-  /// The AdditiveArithmetic protocol in the standard library.
-  ProtocolDecl *additiveArithmeticProtocol =
-      astCtx.getProtocol(KnownProtocolKind::AdditiveArithmetic);
-
-  /// `AdditiveArithmetic.+` declaration.
-  mutable FuncDecl *cachedPlusFn = nullptr;
-  /// `AdditiveArithmetic.+=` declaration.
-  mutable FuncDecl *cachedPlusEqualFn = nullptr;
-
-public:
-  /// Construct an ADContext for the given module.
-  explicit ADContext(SILModuleTransform &transform);
-
-  //--------------------------------------------------------------------------//
-  // General utilities
-  //--------------------------------------------------------------------------//
-
-  SILModuleTransform &getTransform() const { return transform; }
-  SILModule &getModule() const { return module; }
-  ASTContext &getASTContext() const { return module.getASTContext(); }
-  SILPassManager &getPassManager() const { return passManager; }
-  Lowering::TypeConverter &getTypeConverter() { return module.Types; }
-
-  /// Returns true if the `differentiable_function` instruction worklist is
-  /// empty.
-  bool isDifferentiableFunctionInstsWorklistEmpty() const {
-    return differentiableFunctionInsts.empty();
-  }
-
-  /// Pops and returns a `differentiable_function` instruction from the
-  /// worklist. Returns nullptr if the worklist is empty.
-  DifferentiableFunctionInst* popDifferentiableFunctionInstFromWorklist() {
-    if (differentiableFunctionInsts.empty()) return nullptr;
-    return differentiableFunctionInsts.pop_back_val();
-  }
-
-  /// Adds the given `differentiable_function` instruction to the worklist.
-  void addDifferentiableFunctionInst(DifferentiableFunctionInst* dfi) {
-    differentiableFunctionInsts.push_back(dfi);
-  }
-
-  /// Returns true if the given `differentiable_function` instruction has
-  /// already been processed.
-  bool
-  isDifferentiableFunctionInstProcessed(DifferentiableFunctionInst *dfi) const {
-    return processedDifferentiableFunctionInsts.count(dfi);
-  }
-
-  /// Adds the given `differentiable_function` instruction to the worklist.
-  void
-  markDifferentiableFunctionInstAsProcessed(DifferentiableFunctionInst *dfi) {
-    processedDifferentiableFunctionInsts.insert(dfi);
-  }
-
-  const llvm::SmallMapVector<SILDifferentiabilityWitness *,
-                             DifferentiationInvoker, 32> &
-  getInvokers() const {
-    return invokers;
-  }
-
-  void addInvoker(SILDifferentiabilityWitness *witness) {
-    assert(!invokers.count(witness) &&
-           "Differentiability witness already has an invoker");
-    invokers.insert({witness, DifferentiationInvoker(witness)});
-  }
-
-  /// Returns the result index for `dfi` if found in this context. Otherwise,
-  /// sets the result index to zero and returns it.
-  unsigned getResultIndex(DifferentiableFunctionInst *dfi) {
-    return resultIndices[dfi];
-  }
-
-  /// Sets the result index for `dfi`.
-  void setResultIndex(DifferentiableFunctionInst *dfi, unsigned index) {
-    resultIndices[dfi] = index;
-  }
-
-  DenseMap<ApplyInst *, NestedApplyInfo> &getNestedApplyInfo() {
-    return nestedApplyInfo;
-  }
-
-  void recordGeneratedFunction(SILFunction *function) {
-    generatedFunctions.push_back(function);
-  }
-
-  void recordGeneratedFunctionReference(SILValue functionRef) {
-    generatedFunctionReferences.push_back(functionRef);
-  }
-
-  ProtocolDecl *getAdditiveArithmeticProtocol() const {
-    return additiveArithmeticProtocol;
-  }
-
-  FuncDecl *getPlusDecl() const {
-    if (!cachedPlusFn) {
-      cachedPlusFn = findOperatorDeclInProtocol(
-          astCtx.getIdentifier("+"), additiveArithmeticProtocol);
-      assert(cachedPlusFn && "AdditiveArithmetic.+ not found");
-    }
-    return cachedPlusFn;
-  }
-
-  FuncDecl *getPlusEqualDecl() const {
-    if (!cachedPlusEqualFn) {
-      cachedPlusEqualFn = findOperatorDeclInProtocol(
-          astCtx.getIdentifier("+="), additiveArithmeticProtocol);
-      assert(cachedPlusEqualFn && "AdditiveArithmetic.+= not found");
-    }
-    return cachedPlusEqualFn;
-  }
-
-  void cleanUp() {
-    // Delete all references to generated functions.
-    for (auto fnRef : generatedFunctionReferences) {
-      if (auto *fnRefInst =
-              peerThroughFunctionConversions<FunctionRefInst>(fnRef)) {
-        fnRefInst->replaceAllUsesWithUndef();
-        fnRefInst->eraseFromParent();
-      }
-    }
-    // Delete all generated functions.
-    for (auto *generatedFunction : generatedFunctions) {
-      LLVM_DEBUG(getADDebugStream()
-                 << "Deleting generated function "
-                 << generatedFunction->getName() << '\n');
-      generatedFunction->dropAllReferences();
-      transform.notifyWillDeleteFunction(generatedFunction);
-      module.eraseFunction(generatedFunction);
-    }
-  }
-
-  /// Creates an `differentiable_function` instruction using the given builder
-  /// and arguments. Erase the newly created instruction from the processed set,
-  /// if it exists - it may exist in the processed set if it has the same
-  /// pointer value as a previously processed and deleted instruction.
-  /// TODO(TF-784): The pointer reuse is a real concern and the use of
-  /// `CanonicalizeInstruction` may get rid of the need for this workaround.
-  DifferentiableFunctionInst *createDifferentiableFunction(
-      SILBuilder &builder, SILLocation loc,
-      IndexSubset *parameterIndices, SILValue original,
-      Optional<std::pair<SILValue, SILValue>> derivativeFunctions = None) {
-    auto *dfi = builder.createDifferentiableFunction(
-        loc, parameterIndices, original, derivativeFunctions);
-    processedDifferentiableFunctionInsts.erase(dfi);
-    return dfi;
-  }
-
-public:
-  template <typename ...T, typename ...U>
-  InFlightDiagnostic diagnose(SourceLoc loc, Diag<T...> diag,
-                              U &&...args) const {
-    return getASTContext().Diags.diagnose(loc, diag, std::forward<U>(args)...);
-  }
-
-  /// Given an instruction and a differentiation task associated with the
-  /// parent function, emits a "not differentiable" error based on the task. If
-  /// the task is indirect, emits notes all the way up to the outermost task,
-  /// and emits an error at the outer task. Otherwise, emits an error directly.
-  template<typename ...T, typename ...U>
-  InFlightDiagnostic emitNondifferentiabilityError(
-      SILInstruction *inst, DifferentiationInvoker invoker,
-      Diag<T...> diag, U &&...args);
-
-  /// Given a value and a differentiation task associated with the parent
-  /// function, emits a "not differentiable" error based on the task. If the
-  /// task is indirect, emits notes all the way up to the outermost task, and
-  /// emits an error at the outer task. Otherwise, emits an error directly.
-  template<typename ...T, typename ...U>
-  InFlightDiagnostic emitNondifferentiabilityError(
-      SILValue value, DifferentiationInvoker invoker,
-      Diag<T...> diag, U &&...args);
-
-  /// Emit a "not differentiable" error based on the given differentiation task
-  /// and diagnostic.
-  template<typename ...T, typename ...U>
-  InFlightDiagnostic emitNondifferentiabilityError(
-      SourceLoc loc, DifferentiationInvoker invoker,
-      Diag<T...> diag, U &&...args);
-};
-
 class DifferentiationTransformer {
 private:
   /// Reference to the main transform.
@@ -945,107 +658,6 @@ public:
 
 };
 } // end anonymous namespace
-
-ADContext::ADContext(SILModuleTransform &transform)
-    : transform(transform), module(*transform.getModule()),
-      passManager(*transform.getPassManager()) {}
-
-template<typename ...T, typename ...U>
-InFlightDiagnostic
-ADContext::emitNondifferentiabilityError(SILValue value,
-                                         DifferentiationInvoker invoker,
-                                         Diag<T...> diag, U &&...args) {
-  LLVM_DEBUG({
-    getADDebugStream() << "Diagnosing non-differentiability.\n";
-    getADDebugStream() << "For value:\n" << value;
-    getADDebugStream() << "With invoker:\n" << invoker << '\n';
-  });
-  auto valueLoc = value.getLoc().getSourceLoc();
-  // If instruction does not have a valid location, use the function location
-  // as a fallback. Improves diagnostics in some cases.
-  if (valueLoc.isInvalid())
-    valueLoc = value->getFunction()->getLocation().getSourceLoc();
-  return emitNondifferentiabilityError(valueLoc, invoker, diag,
-                                       std::forward<U>(args)...);
-}
-
-template<typename ...T, typename ...U>
-InFlightDiagnostic
-ADContext::emitNondifferentiabilityError(SILInstruction *inst,
-                                         DifferentiationInvoker invoker,
-                                         Diag<T...> diag, U &&...args) {
-  LLVM_DEBUG({
-    getADDebugStream() << "Diagnosing non-differentiability.\n";
-    getADDebugStream() << "For instruction:\n" << *inst;
-    getADDebugStream() << "With invoker:\n" << invoker << '\n';
-  });
-  auto instLoc = inst->getLoc().getSourceLoc();
-  // If instruction does not have a valid location, use the function location
-  // as a fallback. Improves diagnostics for `ref_element_addr` generated in
-  // synthesized stored property getters.
-  if (instLoc.isInvalid())
-    instLoc = inst->getFunction()->getLocation().getSourceLoc();
-  return emitNondifferentiabilityError(instLoc, invoker, diag,
-                                       std::forward<U>(args)...);
-}
-
-template<typename ...T, typename ...U>
-InFlightDiagnostic
-ADContext::emitNondifferentiabilityError(SourceLoc loc,
-                                         DifferentiationInvoker invoker,
-                                         Diag<T...> diag, U &&...args) {
-  switch (invoker.getKind()) {
-  // For `differentiable_function` instructions: if the `differentiable_function`
-  // instruction comes from a differential operator, emit an error on the
-  // expression and a note on the non-differentiable operation. Otherwise, emit
-  // both an error and note on the non-differentiation operation.
-  case DifferentiationInvoker::Kind::DifferentiableFunctionInst: {
-    auto *inst = invoker.getDifferentiableFunctionInst();
-    if (auto *expr = findDifferentialOperator(inst)) {
-      diagnose(expr->getLoc(), diag::autodiff_function_not_differentiable_error)
-          .highlight(expr->getSubExpr()->getSourceRange());
-      return diagnose(loc, diag, std::forward<U>(args)...);
-    }
-    diagnose(loc, diag::autodiff_expression_not_differentiable_error);
-    return diagnose(loc, diag, std::forward<U>(args)...);
-  }
-
-  // For `SILDifferentiabilityWitness`es, try to find an AST function
-  // declaration and `@differentiable` attribute. If they are found, emit an
-  // error on the `@differentiable` attribute; otherwise, emit an error on the
-  // SIL function. Emit a note at the non-differentiable operation.
-  case DifferentiationInvoker::Kind::SILDifferentiabilityWitnessInvoker: {
-    auto *witness = invoker.getSILDifferentiabilityWitnessInvoker();
-    auto *original = witness->getOriginalFunction();
-    if (auto *diffAttr = witness->getAttribute()) {
-      diagnose(diffAttr->getLocation(),
-               diag::autodiff_function_not_differentiable_error)
-          .highlight(diffAttr->getRangeWithAt());
-      diagnose(original->getLocation().getSourceLoc(),
-               diag::autodiff_when_differentiating_function_definition);
-    } else {
-      diagnose(original->getLocation().getSourceLoc(),
-               diag::autodiff_function_not_differentiable_error);
-    }
-    return diagnose(loc, diag, std::forward<U>(args)...);
-  }
-
-  // For indirect differentiation, emit a "not differentiable" note on the
-  // expression first. Then emit an error at the source invoker of
-  // differentiation, and a "when differentiating this" note at each indirect
-  // invoker.
-  case DifferentiationInvoker::Kind::IndirectDifferentiation: {
-    SILInstruction *inst;
-    SILDifferentiabilityWitness *witness;
-    std::tie(inst, witness) = invoker.getIndirectDifferentiation();
-    auto invokerLookup = invokers.find(witness);
-    assert(invokerLookup != invokers.end() && "Expected parent invoker");
-    emitNondifferentiabilityError(inst, invokerLookup->second,
-        diag::autodiff_expression_not_differentiable_note);
-    return diagnose(loc, diag::autodiff_when_differentiating_function_call);
-  }
-  }
-}
 
 
 namespace {
