+++ lib/SILOptimizer/Utils/Differentiation/CMakeLists.txt
@@ -1,4 +1,5 @@
 silopt_register_sources(
+  ADContext.cpp
   Common.cpp
   DerivativeLookup.cpp
   DifferentiationInvoker.cpp
